refactor: rename delegate-monitor components
