+++ src/components/monitor/header/index.js
@@ -1,0 +1,9 @@
+import DelegateCount from './DelegateCount'
+import LastBlock from './LastBlock'
+import TotalForged from './TotalForged'
+
+export {
+  DelegateCount,
+  LastBlock,
+  TotalForged
+}
