+++ src/components/monitor/index.js
@@ -1,0 +1,13 @@
+import ActiveDelegates from './ActiveDelegates'
+import MonitorHeader from './MonitorHeader'
+import ForgingStats from './ForgingStats'
+import StandbyDelegates from './StandbyDelegates'
+
+export * from './header'
+
+export {
+  ActiveDelegates,
+  MonitorHeader,
+  ForgingStats,
+  StandbyDelegates
+}
